Under Graduate 8th Semester Project
Abstract- The technique of recognizing an item or 
feature in an image or video is known as image 
recognition. It is employed in several contexts, 
including flaw identification, imaging in the 
medical field, and security monitoring. Most 
vision-based AI systems and applications today 
are focused on object recognition. Object 
recognition is crucial for clear determination, 
which is helpful in security, transportation, 
healthcare, and military use cases object 
detection in retail. Due to the large range of dress 
designs and the intricate nature of their 
commonalities, it is difficult to precisely 
categorize police uniforms. In Bangladesh, 
uniforms for the police and security forces come 
in a variety of hues. This makes it difficult to 
improve performance using a simple 
classification system. However, using this 
strategy will ensure that you may conduct study 
in this exciting field. Consequently, here is the 
technique for identifying Bangladeshi police 
uniforms from a picture was developed. In order 
to achieve these goals, we first constructed the 
CNN model for predicting Bangla digits since 
CNN is the most intelligent object prediction 
model. After a thorough examination, the 
suggested system attained a better accuracy on 
our own customized dataset.